requirements.txt
django==3.2.6
pillow==8.2.0
django-environ==0.4.5
gunicorn==20.1.0
whitenoise==5.2.0
django-import-export==2.7.1
pipwin==0.5.1
PyMySQL==1.0.2
django-mathfilters==1.0.0